requirements.txt
asgiref==3.7.2
Babel==2.13.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
Django==4.2.3
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-js-asset==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
forex-python==1.8
h5py==3.9.0
idna==3.4
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
numpy==1.25.2
Pillow==10.0.0
# psycopg2==2.9.6
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.1
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4